+++ integrations/discord/events/comparison.py
@@ -10,7 +10,8 @@ from discord import Message
 from discord.channel import TextChannel
 
 import libs.global_value as g
-from libs.data import modify, search
+from libs.data import search
+from libs.domain import modify
 from libs.domain.datamodels import ComparisonResults
 from libs.domain.score import GameResult
 from libs.types import ActionStatus, CommandType, RemarkDict, StyleOptions
