+++ integrations/slack/events/comparison.py
@@ -6,7 +6,8 @@ import logging
 from typing import TYPE_CHECKING, cast
 
 import libs.global_value as g
-from libs.data import modify, search
+from libs.data import search
+from libs.domain import modify
 from libs.domain.datamodels import ComparisonResults
 from libs.domain.score import GameResult
 from libs.functions import lookup
