+++ libs/data/__init__.py
@@ -1,6 +1,5 @@
 """
 データ操作モジュール
 
-- `libs.data.modify`: DBデータ操作
 - `libs.data.search`: DB検索
 """
