+++ libs/domain/__init__.py
@@ -8,4 +8,5 @@
 - `libs.domain.rule`: ルールセット管理
 - `libs.domain.placeholder`: プレースホルダ構築
 - `libs.domain.aggregate`: データ集計
+- `libs.domain.modify`: DBデータ操作
 """
