+++ libs/functions/tools/recalculation.py
@@ -6,7 +6,7 @@ import logging
 from contextlib import closing
 
 import libs.global_value as g
-from libs.data import modify
+from libs.domain import modify
 from libs.domain.score import GameResult
 from libs.utils import dbutil, dictutil
