+++ tests/database/test_result_update.py
@@ -10,7 +10,7 @@ import pytest
 import libs.global_value as g
 from integrations import factory
 from libs.bootstrap import configuration, initialization
-from libs.data import modify
+from libs.domain import modify
 from libs.domain.score import GameResult
 from libs.types import ServiceType
 from libs.utils import dbutil, validator
